cleanup(core): remove the indirection left after inlining tao into nx
