+++ .circleci/config.yml
@@ -188,7 +188,7 @@ jobs:
       - run:
           name: Run E2E Tests
           command: |
-            npx nx affected --target=e2e --base=$NX_BASE --head=$NX_HEAD --exclude=e2e-storybook,e2e-react-native,e2e-detox --parallel=1
+            npx nx affected --target=e2e --base=$NX_BASE --head=$NX_HEAD --exclude=e2e-storybook,e2e-react-native,e2e-detox,e2e-workspace-integrations --parallel=1
           no_output_timeout: 45m
       - run:
           name: Stop All Running Agents for This CI Run
