+++ packages/nx-plugin/src/generators/migration/schema.json
@@ -46,6 +46,6 @@
       "default": false
     }
   },
-  "required": ["project", "version"],
+  "required": ["project", "packageVersion"],
   "additionalProperties": false
 }
